requirements.txt
beautifulsoup4>=4.11.0
selenium>=4.10.0
requests>=2.28.0
lxml>=4.9.0
python-docx>=0.8.11
fake-useragent>=1.4.0
streamlit>=1.28.0
pytest>=7.4.0
pytest-mock>=3.11.0
openpyxl>=3.1.0
